2020 11

+++ puzzles/2022/11/src/types.rs
@@ -1,1 +1,1 @@
-pub type WorryLevel = u32;
+pub type WorryLevel = u64;
